Make c10::Device::validate() debug-only assert.
Callgrind profiling suggests that c10::Device::validate() is
a hot spot on the "add tensors together microbenchmark".  In
normal PyTorch use, validation of device arguments is done
in the Python binding level, which suggests we should just
assume that C++ clients have validated everything already.
Switch this to debug-only assert.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: dd2c3f6381e56d4587314d32bcd4f264543d5d1d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44322