[NCCL] Add Environment Variable to guard Async Error Handling feature
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44163

In this PR, we introduce a new environment variable
(NCCL_ASYNC_ERROR_HANDLING), which guards the asynchronous error handling
feature. We intend to eventually turn this feature on by default for all users,
but this is a temporary solution so the change in behavior from hanging to
crashing is not the default for users all of a sudden.
ghstack-source-id: 111614317

Differential Revision: [D23517895](https://our.internmc.facebook.com/intern/diff/D23517895/)